REF: Move callbacks into separate module
